fnhw09-01-fa-dev-bbk/requirements.txt
# Do not include azure-functions-worker as it may conflict with the Azure Functions platform

azure-functions==1.7.2
azure-identity==1.6.0
azure-mgmt-keyvault==9.0.0
azure-mgmt-resource==19.0.0